requirements.txt
asgiref==3.4.1
autopep8==1.5.7
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.6
django-cors-headers==3.8.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
idna==3.2
psycopg2-binary==2.9.1
pycodestyle==2.7.0
PyJWT==2.1.0
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
sentry-sdk==1.3.1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.6
